frameworks/crewai/unit-test-agent-demo/pyproject.toml
[project]
name = "unit-test-agent-demo"
version = "0.1.0"
description = "CrewAI-based Unit Test Agent that automatically generates test plans and code for Python functions"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "crewai>=0.86.0",
    "crewai-tools>=0.17.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "gitpython>=3.1.0",
    "astroid>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.14.0",
    "black>=24.0.0",
    "ruff>=0.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src --cov-report=term-missing"
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "scenario: Scenario-based tests",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.black]
line-length = 100
target-version = ["py311"]
